magento/adobe-stock-integration#3: Refactored API structure
