+++ AdobeStockAsset/Model/Request/Config/Converter.php
@@ -3,7 +3,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\AdobeStockAsset\Model\Request\Builder\Config;
+namespace Magento\AdobeStockAsset\Model\Request\Config;
 
 class Converter implements \Magento\Framework\Config\ConverterInterface
 {
