+++ AdobeStockAsset/Model/Request/Config/Reader.php
@@ -3,7 +3,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\AdobeStockAsset\Model\Request\Builder\Config;
+namespace Magento\AdobeStockAsset\Model\Request\Config;
 
 class Reader extends \Magento\Framework\Config\Reader\Filesystem
 {
