+++ AdobeStockAsset/Model/Request/ConfigInterface.php
@@ -3,7 +3,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\AdobeStockAsset\Model\Request\Builder;
+namespace Magento\AdobeStockAsset\Model\Request;
 
 /**
  * Configured postcode validation patterns
