+++ AdobeStockAssetApi/Api/SearchRequestBuilderInterface.php
@@ -10,7 +10,7 @@ namespace Magento\AdobeStockAssetApi\Api;
  * Interface RequestBuilderInterface
  * @package Magento\AdobeStockAssetApi\Api
  */
-interface RequestBuilderInterface
+interface SearchRequestBuilderInterface
 {
     /**
      * Set request name
@@ -61,7 +61,7 @@ interface RequestBuilderInterface
     public function bind(string $name, $value) : void;
 
     /**
-     * @return \Magento\AdobeStockAssetApi\Api\Data\RequestInterface
+     * @return \Magento\AdobeStockAssetApi\Api\Data\SearchRequestInterface
      */
-    public function create() : \Magento\AdobeStockAssetApi\Api\Data\RequestInterface;
+    public function create() : \Magento\AdobeStockAssetApi\Api\Data\SearchRequestInterface;
 }
